Abstract search bar and render on different routes
